pyproject.toml
[tool.poetry]
name = "project_x"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Your Name <you@example.com>"]
readme = 'README.md'
repository = "https://github.com/sudomaze/project_x"
homepage = "https://github.com/sudomaze/project_x"
keywords = ['pytorch', 'pytorch-lightning']

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sudomaze/project_x/issues"
"Changelog" = "https://github.com/sudomaze/project_x/blob/master/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = {url = "https://download.pytorch.org/whl/cu111/torch-1.9.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu111/torchvision-0.10.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}
torchaudio = {url = "https://download.pytorch.org/whl/torchaudio-0.9.0-cp39-cp39-linux_x86_64.whl"}
cuda-slic = "^0.0.1-alpha.3"
pytorch-lightning = "^1.4.5"
PyYAML = "^5.4.1"
fire = "^0.4.0"
tensorboard = "^2.6.0"
jupyter = "^1.0.0"
albumentations = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pylint = "^2.10.2"
autopep8 = "^1.5.7"
visdom = "^0.1.8"
jsonpatch = "^1.32"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120 # override the default of 88

[tool.poetry.scripts]
train = "scripts.train:train"

# [tool.poe.tasks]
# test = { "cmd" = "pytest tests --cov=. --cov-report term-missing", "help" = "Run all tests and get Code Coverage Report" }
# test-postman = { "cmd" = "pytest tests/test_postman.py", "help" = "Run only tests against Postman" }
# lint = [{ "ref" = "_lint-syntax" }, { "ref" = "_lint-full" }]

# # private tasks
# _lint-syntax = { "cmd" = "flake8 ./pyclinic ./tests --select=E9,F63,F7,F82 --show-source", "help" = "Lint for syntax error and undefined names" }
# _lint-full = { "cmd" = "flake8 ./pyclinic ./tests --exit-zero --max-complexity=10", "help" = "Lint for style, complexity, and everything else" }


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
